Refactor the space_after / misc code to models.common.utils so it can be used elsewhere
